Linear Regression refactor #15
