+++ crates/rustlearn-linear-model/src/ols/mod.rs
@@ -1,2 +1,1 @@
 pub mod linear_regression;
-pub mod multiple_regression;
