+++ py-rustlearn/src/lib.rs
@@ -1,8 +1,7 @@
 use pyo3::prelude::*;
 use rustlearn_python::array::PyNamedArray;
 use rustlearn_python::exceptions::{NotYetImplementedError, RustLearnError, ValidationError};
-use rustlearn_python::linear_model::linear_regression::PySimpleLinearRegression;
-use rustlearn_python::linear_model::multiple_regression::PyMultipleLinearRegression;
+use rustlearn_python::linear_model::linear_regression::PyLinearRegression;
 use rustlearn_python::linear_model::PyLinearRegressionReturn;
 use rustlearn_python::metrics::{mean_absolute_error, mean_squared_error};
 
@@ -11,8 +10,7 @@ fn rustylearn(py: Python, m: &Bound<PyModule>) -> PyResult<()> {
     // Base Classes
     m.add_class::<PyNamedArray>().unwrap();
     m.add_class::<PyLinearRegressionReturn>().unwrap();
-    m.add_class::<PySimpleLinearRegression>().unwrap();
-    m.add_class::<PyMultipleLinearRegression>().unwrap();
+    m.add_class::<PyLinearRegression>().unwrap();
 
     // Metrics
     m.add_wrapped(wrap_pyfunction!(
